Documents/src/challenges/ctf-stuff/AI CTF/playing_with_fonts/fontplay/pyproject.toml
[tool.poetry]
name = "fontplay"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.28.1"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
fastapi = "^0.111.0"
jinja2 = "^3.1.4"
uvicorn = "^0.29.0"
starlette-wtf = "^0.4.5"
itsdangerous = "^2.2.0"
httpx = "^0.27.0"
pydub = "^0.25.1"

[tool.poetry.scripts]
fontplay = "fontplay:run"
fontplay-debug = "fontplay:debug"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
